Moving Fermi to Spec
Since there is only one script that makes the fermi analysis for the
FOXSI-1 flight, we moved it to the spec folder. The interested user can
use this script as an example to analyse FOXSI-2. In case we get a
critical amount of scripts for Fermi analysis, we will consider to make
an exclusive folder for this kind of analysis.