Image routines update
This is an update of the way images are created for the FOXSI data. Silicon detector are not impacted and the updates should not change the way images are done with Silicon detectors. The changes for the CdTe are the following: a different definition of the detector coordinates in level 1 data, a different call to get_payload_coordinates which does not require any reflection of the y axis but does require a rotation with the opposite angle of what is used for Silicon. The solar coordinates in lvl2 are calculated with a possible offset and roll angle. The foxsi_image_det is updated to take CdTe as a keyword. The foxsi_image_map is updated to work with both Silicon and CdTe, with the provided keyword.