Create new files to be called in FOXSI.pro
undefine_previous_param is used to delete variables than can already exist and avoid confusion. For example, if I first worked with foxsi2 data, I have the variable cen0_pos0 defined. If I switch to FOXSI3 data, the variable cen0 will be created, but the variable cen0_pos0 does not exists for FOXSI3, and so if we do not delete the variable it will still exists. If now I am using a piece of code calling cen0_pos0, it will not produce any error message and continue with the position information from FOXSI2, which is not what is expected. That is why we need to delete possible variables before initializing them.
The file param2018_201910015 contains parameters for FOXSI3