Complete Effective Area for Modules X7 and X8 which are the two new 10-Shell Modules
I included a total of four files which contain the Effective Area for the new 10-shell modules flown in FOXSI-3. The files are written following the same format used for FOXSI-2. The pan and tilt off-axis angles were different from the ones used in previous flights. This, requiere the addition of some of the scripts in this repo that call the files to account for these new restrictions.

We need to discuss whether or not to include files for the Effective Area for all the other modules flown in FOXSI-3. During the last set of calibrations, there were limited number of measurements for the EA at off-axis angles. Also, we need to assess if the EA changed from FOXSI-2 to FOXSI-3 for those optics that were flown in both rockets.