Update resp/get_foxsi_optics_effarea
I included the COMMON FOXSI set of parameters to have access to the DATE of launch. That way we can differenciate among foxsi 1, 2 and 3 to get the EffArea.

I implemented some CASES sorted by DATE.

I changed the way the DATA struct is created to parse the dimensions from the calibration files themselves.

To test the code run SSWIDL. Then
> foxsi, 2012
or
> foxsi, 2014
then
> area = get_foxsi_optics_effarea( module_number=0,offaxis_angle=offaxis_angle, data_dir=data_dir)

This should work for foxsi1 and foxsi2. in the case of foxsi3, the module_number should be 7 or 8. However, I still have some problems due to pan and tilt having different off-axis angles into the calibration files themselves. Further work is needed in here to successfully implement this code for FOXSI3.