Answering Sophie comments and fixing EA for FOXSI3
I addressed all the comments Sophie gave me this morning. In particular, I responded to te question of why to use the DATE instead of the YEAR variable. Also I switched from CASE to IFs and ELSEs to acoount for the module numbers corresponding to each foxsi flight. I finally updated the FOXSI3 X7 and X8 tilt files to fix the bug of the code when running FOXSI3.