Load parameters when setting /read_fits
I was testing this code with the FOXSI fits files that are posted on the VSO and noticed that the parameters weren't being loaded when the keyword "read_fits" was set. To remedy this, I included the call to load the appropriate parameters in the "case" statement starting on line 53.