Included the YEAR variable
get_foxsi_optics_effarea now includes an optional YEAR variable that if set there is not need of loading the whole FOXSI COMMON variables.

If Yar is not set, then the user needs to load FOXSI, YEAR.

If none of this two options are slected, a warning message shows up to the user.